+++ crates/boatramp-handlers/src/engine.rs
@@ -451,7 +451,7 @@ impl HandlerEngine {
         })?;
         #[cfg(feature = "invoke")]
         bindings::invoke::add_to_linker(&mut linker, |state: &mut HostState| {
-            bindings::invoke::InvokeHost::new(state.bindings.invoke())
+            bindings::invoke::InvokeHost::new(&mut state.table, state.bindings.invoke())
         })?;
         Ok(linker)
     }